+++ src/main/java/racingcar/Model/Car.java
@@ -1,9 +1,11 @@
-package racingcar.Model;
+package racingcar.model;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import racingcar.Enum.TypeOfLocation;
+
+import racingcar.enums.TypeOfLocation;
+
 
 public class Car {
   private static final String NAME_AND_LOCATION_DELIMITER = " : ";
