+++ src/main/java/racingcar/Model/RacingRule.java
@@ -1,4 +1,4 @@
-package racingcar.Model;
+package racingcar.model;
 import java.util.ArrayList;
 
 public interface RacingRule {
