+++ src/main/java/racingcar/Utilities/Random.java
@@ -1,4 +1,4 @@
-package racingcar.Utilities;
+package racingcar.utilities;
 
 import camp.nextstep.edu.missionutils.Randoms;
