+++ src/main/java/racingcar/enums/IOMessage.java
@@ -1,4 +1,4 @@
-package racingcar.Enum;
+package racingcar.enums;
 
 public enum IOMessage implements SystemMessage {
   INPUT_CARS_NAME("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
