+++ src/main/java/racingcar/enums/SystemMessage.java
@@ -1,4 +1,4 @@
-package racingcar.Enum;
+package racingcar.enums;
 
 public interface SystemMessage {
   String getMessage();
