+++ src/test/java/racingcar/ApplicationTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
-import racingcar.Enum.ErrorMessage;
+import racingcar.enums.ErrorMessage;
 
 import org.junit.jupiter.api.Test;
