+++ src/test/java/racingcar/CarValidatorTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import racingcar.Enum.ErrorMessage;
-import racingcar.Validation.CarValidator;
+import racingcar.enums.ErrorMessage;
+import racingcar.validation.CarValidator;
 
 public class CarValidatorTest {
