+++ src/test/java/racingcar/RacingTest.java
@@ -5,8 +5,9 @@ import static org.assertj.core.api.Assertions.*;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import racingcar.Model.Car;
-import racingcar.Model.Racing;
+
+import racingcar.model.Car;
+import racingcar.model.Racing;
 
 public class RacingTest {
